With a lot of bugs.trying to fix them!
有很多bug需要改！

+++ Assets/Photon Unity Networking/Plugins/PhotonNetwork/PhotonView.cs
@@ -64,7 +64,7 @@ public class PhotonView : Photon.MonoBehaviour
 
     public int ownerId;
 
-    public int group = 0;
+    public byte group = 0;
 
     protected internal bool mixedModeIsReliable = false;
